refactor(v2): renamed and moved integration tests into their own folder
