+++ packages/qwik/src/core/v2/tests/use-computed.test.tsx
@@ -1,17 +1,17 @@
-import { describe, expect, it } from 'vitest';
-import { trigger } from '../../testing/element-fixture';
 import {
-  component$,
-  useSignal,
-  useStore,
-  useComputed$,
   Fragment as Signal,
+  component$,
+  qrl,
+  useComputed$,
   useComputedQrl,
   useLexicalScope,
-  qrl,
+  useSignal,
+  useStore,
 } from '@builder.io/qwik';
-import { domRender, ssrRenderToDom } from '../../testing/rendering.unit-util';
-import '../../testing/vdom-diff.unit-util';
+import { describe, expect, it } from 'vitest';
+import { trigger } from '../../../testing/element-fixture';
+import { domRender, ssrRenderToDom } from '../../../testing/rendering.unit-util';
+import '../../../testing/vdom-diff.unit-util';
 
 const debug = false; //true;
 Error.stackTraceLimit = 100;
