+++ packages/qwik/src/core/v2/tests/use-on.test.tsx
@@ -1,5 +1,3 @@
-import { describe, expect, it } from 'vitest';
-import { trigger } from '../../testing/element-fixture';
 import {
   $,
   Fragment as Component,
@@ -11,8 +9,10 @@ import {
   useSignal,
   useVisibleTask$,
 } from '@builder.io/qwik';
-import { domRender, ssrRenderToDom } from '../../testing/rendering.unit-util';
-import '../../testing/vdom-diff.unit-util';
+import { describe, expect, it } from 'vitest';
+import { trigger } from '../../../testing/element-fixture';
+import { domRender, ssrRenderToDom } from '../../../testing/rendering.unit-util';
+import '../../../testing/vdom-diff.unit-util';
 
 const debug = false; //true;
 Error.stackTraceLimit = 100;
