+++ packages/qwik/src/core/v2/tests/use-sequential-scope.test.tsx
@@ -1,9 +1,9 @@
 import { describe, expect, it } from 'vitest';
-import { trigger } from '../../testing/element-fixture';
-import { component$ } from '../component/component.public';
-import { useSequentialScope } from '../use/use-sequential-scope';
-import { domRender, rerenderComponent, ssrRenderToDom } from '../../testing/rendering.unit-util';
-import '../../testing/vdom-diff.unit-util';
+import { trigger } from '../../../testing/element-fixture';
+import { domRender, rerenderComponent, ssrRenderToDom } from '../../../testing/rendering.unit-util';
+import '../../../testing/vdom-diff.unit-util';
+import { component$ } from '../../component/component.public';
+import { useSequentialScope } from '../../use/use-sequential-scope';
 
 const debug = false; //true;
 Error.stackTraceLimit = 100;
