+++ packages/qwik/src/core/v2/tests/use-signal.test.tsx
@@ -6,14 +6,14 @@ import {
   Fragment as Signal,
 } from '@builder.io/qwik';
 import { describe, expect, it } from 'vitest';
-import { trigger } from '../../testing/element-fixture';
-import { component$ } from '../component/component.public';
-import { Slot } from '../render/jsx/slot.public';
-import type { Signal as SignalType } from '../state/signal';
-import { untrack } from '../use/use-core';
-import { useSignal } from '../use/use-signal';
-import { domRender, ssrRenderToDom } from '../../testing/rendering.unit-util';
-import '../../testing/vdom-diff.unit-util';
+import { trigger } from '../../../testing/element-fixture';
+import { domRender, ssrRenderToDom } from '../../../testing/rendering.unit-util';
+import '../../../testing/vdom-diff.unit-util';
+import { component$ } from '../../component/component.public';
+import { Slot } from '../../render/jsx/slot.public';
+import type { Signal as SignalType } from '../../state/signal';
+import { untrack } from '../../use/use-core';
+import { useSignal } from '../../use/use-signal';
 
 const debug = false; //true;
 Error.stackTraceLimit = 100;
