+++ packages/qwik/src/core/v2/tests/use-styles-scoped.test.tsx
@@ -1,24 +1,24 @@
-import { createDocument } from '@builder.io/qwik-dom';
 import {
   Fragment as Component,
   Fragment,
   Fragment as Projection,
   Fragment as Signal,
 } from '@builder.io/qwik';
+import { createDocument } from '@builder.io/qwik-dom';
 import { afterEach, describe, expect, it } from 'vitest';
-import { useStore } from '..';
-import { renderToString2 } from '../../server/v2-ssr-render2';
-import { trigger } from '../../testing/element-fixture';
-import { component$ } from '../component/component.public';
-import { getPlatform, setPlatform } from '../platform/platform';
-import { inlinedQrl } from '../qrl/qrl';
-import { Slot } from '../render/jsx/slot.public';
-import { getScopedStyles } from '../style/scoped-stylesheet';
-import { useSignal } from '../use/use-signal';
-import { useStylesScopedQrl } from '../use/use-styles';
-import { QStyleSelector } from '../util/markers';
-import { domRender, ssrRenderToDom } from '../../testing/rendering.unit-util';
-import '../../testing/vdom-diff.unit-util';
+import { useStore } from '../..';
+import { renderToString2 } from '../../../server/v2-ssr-render2';
+import { trigger } from '../../../testing/element-fixture';
+import { domRender, ssrRenderToDom } from '../../../testing/rendering.unit-util';
+import '../../../testing/vdom-diff.unit-util';
+import { component$ } from '../../component/component.public';
+import { getPlatform, setPlatform } from '../../platform/platform';
+import { inlinedQrl } from '../../qrl/qrl';
+import { Slot } from '../../render/jsx/slot.public';
+import { getScopedStyles } from '../../style/scoped-stylesheet';
+import { useSignal } from '../../use/use-signal';
+import { useStylesScopedQrl } from '../../use/use-styles';
+import { QStyleSelector } from '../../util/markers';
 
 const debug = false; //true;
 Error.stackTraceLimit = 100;
