+++ packages/qwik/src/core/v2/tests/use-styles.test.tsx
@@ -1,17 +1,17 @@
 import { Fragment as Component, Fragment as Signal } from '@builder.io/qwik';
-import { afterEach, describe, expect, it } from 'vitest';
-import { trigger } from '../../testing/element-fixture';
-import { component$ } from '../component/component.public';
-import { inlinedQrl } from '../qrl/qrl';
-import { useSignal } from '../use/use-signal';
-import { useStylesQrl } from '../use/use-styles';
-import { domRender, ssrRenderToDom } from '../../testing/rendering.unit-util';
-import '../../testing/vdom-diff.unit-util';
-import { QStyleSelector } from '../util/markers';
-import { renderToString2 } from '../../server/v2-ssr-render2';
-import { Slot } from '../render/jsx/slot.public';
-import { getPlatform, setPlatform } from '../platform/platform';
 import { createDocument } from '@builder.io/qwik-dom';
+import { afterEach, describe, expect, it } from 'vitest';
+import { renderToString2 } from '../../../server/v2-ssr-render2';
+import { trigger } from '../../../testing/element-fixture';
+import { domRender, ssrRenderToDom } from '../../../testing/rendering.unit-util';
+import '../../../testing/vdom-diff.unit-util';
+import { component$ } from '../../component/component.public';
+import { getPlatform, setPlatform } from '../../platform/platform';
+import { inlinedQrl } from '../../qrl/qrl';
+import { Slot } from '../../render/jsx/slot.public';
+import { useSignal } from '../../use/use-signal';
+import { useStylesQrl } from '../../use/use-styles';
+import { QStyleSelector } from '../../util/markers';
 
 const debug = false; //true;
 Error.stackTraceLimit = 100;
