+++ packages/qwik/src/core/v2/tests/use-task.test.tsx
@@ -1,15 +1,15 @@
 import { describe, expect, it } from 'vitest';
-import { trigger } from '../../testing/element-fixture';
-import { component$ } from '../component/component.public';
-import { Fragment, Fragment as Component, Fragment as Signal } from '../render/jsx/jsx-runtime';
-import type { Signal as SignalType } from '../state/signal';
-import { useSignal } from '../use/use-signal';
-import { useStore } from '../use/use-store.public';
-import { useTask$ } from '../use/use-task';
-import { delay } from '../util/promises';
-import { ErrorProvider, domRender, ssrRenderToDom } from '../../testing/rendering.unit-util';
-import '../../testing/vdom-diff.unit-util';
-import { getTestPlatform } from '../../testing/platform';
+import { trigger } from '../../../testing/element-fixture';
+import { getTestPlatform } from '../../../testing/platform';
+import { ErrorProvider, domRender, ssrRenderToDom } from '../../../testing/rendering.unit-util';
+import '../../../testing/vdom-diff.unit-util';
+import { component$ } from '../../component/component.public';
+import { Fragment as Component, Fragment, Fragment as Signal } from '../../render/jsx/jsx-runtime';
+import type { Signal as SignalType } from '../../state/signal';
+import { useSignal } from '../../use/use-signal';
+import { useStore } from '../../use/use-store.public';
+import { useTask$ } from '../../use/use-task';
+import { delay } from '../../util/promises';
 
 const debug = false; //true;
 Error.stackTraceLimit = 100;
