+++ vitest.config.ts
@@ -1,6 +1,6 @@
-import { defineConfig } from 'vitest/config';
 import { qwikVite } from '@builder.io/qwik/optimizer';
 import tsconfigPaths from 'vite-tsconfig-paths';
+import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   plugins: [
@@ -12,6 +12,7 @@ export default defineConfig({
   ],
   test: {
     include: [
+      'packages/**/*.test.?(c|m)[jt]s?(x)',
       'packages/**/*.unit.?(c|m)[jt]s?(x)',
       '!packages/qwik/dist',
       '!packages/*/lib',
